@@ -1,16 +1,18 @@
 ﻿using System;
 using System.IO;
 using System.Reflection;
+using System.Threading.Tasks;
 using IdentityServer4.Contrib.AspNetCore.Testing.Builder;
 using IdentityServer4.Services;
 using IdentityServer4.Testing.Infrastructure.Services;
 using IdentityServer4.Testing.Infrastructure.Validators;
 using IdentityServer4.Validation;
-using Microsoft.AspNetCore.Hosting.Internal;
+using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Serilog;
+using Serilog.Events;
 using Serilog.Sinks.SystemConsole.Themes;
 using Xunit;
 
@@ -21,21 +23,18 @@ namespace IdentityServer4.Contrib.AspNetCore.Testing.Tests
         [Fact]
         public void IdentityServerWebHostBuilder_UseConfigurationBuilder_HasSettings()
         {
-            var webHost = new IdentityServerWebHostBuilder()
+            var webHost = new IdentityServerHostBuilder()
                 .UseConfigurationBuilder((context, builder) =>
                 {
-                    context.HostingEnvironment = new HostingEnvironment
-                    {
-                        ContentRootPath = AppContext.BaseDirectory
-                    };
+                    context.HostingEnvironment.WebRootPath = AppContext.BaseDirectory;
                     builder.AddJsonFile(Path.Combine(AppContext.BaseDirectory, "testappsettings.json"), false);
                 })
-                .CreateWebHostBuilder()
+                .CreateWebHostBuider()
                 .Build();
 
             var configuration = webHost.Services.GetRequiredService<IConfiguration>();
             var hostingEnvironment =
-                webHost.Services.GetRequiredService<Microsoft.Extensions.Hosting.IHostingEnvironment>();
+                webHost.Services.GetRequiredService<IWebHostEnvironment>();
 
             Assert.Equal("PropValue", configuration["Prop"]);
             Assert.Equal(AppContext.BaseDirectory, hostingEnvironment.ContentRootPath);
@@ -44,9 +43,9 @@ namespace IdentityServer4.Contrib.AspNetCore.Testing.Tests
         [Fact]
         public void IdentityServerWebHostBuilder_UseProfileService_Resolveable()
         {
-            var webHost = new IdentityServerWebHostBuilder()
+            var webHost = new IdentityServerHostBuilder()
                 .UseProfileService(new SimpleProfileService())
-                .CreateWebHostBuilder()
+                .CreateWebHostBuider()
                 .Build();
 
             webHost.Services.GetRequiredService<IProfileService>();
@@ -55,9 +54,9 @@ namespace IdentityServer4.Contrib.AspNetCore.Testing.Tests
         [Fact]
         public void IdentityServerWebHostBuilder_UseProfileService_Typed_Resolveable()
         {
-            var webHost = new IdentityServerWebHostBuilder()
+            var webHost = new IdentityServerHostBuilder()
                 .UseProfileService(typeof(SimpleProfileService))
-                .CreateWebHostBuilder()
+                .CreateWebHostBuider()
                 .Build();
 
             webHost.Services.GetRequiredService<IProfileService>();
@@ -66,16 +65,16 @@ namespace IdentityServer4.Contrib.AspNetCore.Testing.Tests
         [Fact]
         public void IdentityServerWebHostBuilder_UseProfileService_Typed_Invalid_Throws()
         {
-            Assert.Throws<ArgumentException>(() => new IdentityServerWebHostBuilder()
+            Assert.Throws<ArgumentException>(() => new IdentityServerHostBuilder()
                 .UseProfileService(typeof(ExtensionsGrantValidator)));
         }
 
         [Fact]
         public void IdentityServerWebHostBuilder_UseResourceOwnerPasswordValidator_Resolveable()
         {
-            var webHost = new IdentityServerWebHostBuilder()
+            var webHost = new IdentityServerHostBuilder()
                 .UseResourceOwnerPasswordValidator(new SimpleResourceOwnerPasswordValidator())
-                .CreateWebHostBuilder()
+                .CreateWebHostBuider()
                 .Build();
 
             webHost.Services.GetRequiredService<IResourceOwnerPasswordValidator>();
@@ -86,10 +85,10 @@ namespace IdentityServer4.Contrib.AspNetCore.Testing.Tests
         {
             InitializeSerilog();
 
-            var webHost = new IdentityServerWebHostBuilder()
+            var webHost = new IdentityServerHostBuilder()
                 .UseLoggingBuilder((context, builder) => builder.AddSerilog())
                 .UseResourceOwnerPasswordValidator(typeof(ResourceOwnerValidatorWithDependencies))
-                .CreateWebHostBuilder()
+                .CreateWebHostBuider()
                 .Build();
 
             webHost.Services.GetRequiredService<IResourceOwnerPasswordValidator>();
@@ -98,9 +97,9 @@ namespace IdentityServer4.Contrib.AspNetCore.Testing.Tests
         [Fact]
         public void IdentityServerWebHostBuilder_UseResourceOwnerPasswordValidator_Typed_Resolveable()
         {
-            var webHost = new IdentityServerWebHostBuilder()
+            var webHost = new IdentityServerHostBuilder()
                 .UseResourceOwnerPasswordValidator(typeof(SimpleResourceOwnerPasswordValidator))
-                .CreateWebHostBuilder()
+                .CreateWebHostBuider()
                 .Build();
 
             webHost.Services.GetRequiredService<IResourceOwnerPasswordValidator>();
@@ -109,18 +108,19 @@ namespace IdentityServer4.Contrib.AspNetCore.Testing.Tests
         [Fact]
         public void IdentityServerWebHostBuilder_UseResourceOwnerPasswordValidator_Typed_Invalid_Throws()
         {
-            Assert.Throws<ArgumentException>(() => new IdentityServerWebHostBuilder()
+            Assert.Throws<ArgumentException>(() => new IdentityServerHostBuilder()
                 .UseResourceOwnerPasswordValidator(typeof(ExtensionsGrantValidator)));
         }
 
         [Fact]
-        public void IdentityServerWebHostBuilder_UseLoggingBuilder_Serilog_Expect_Logger_And_Provider()
+        public async Task IdentityServerWebHostBuilder_UseLoggingBuilder_Serilog_Expect_Logger_And_Provider()
         {
             InitializeSerilog();
 
-            var webHost = new IdentityServerWebHostBuilder()
+            var webHost = new IdentityServerHostBuilder()
                 .UseLoggingBuilder((context, builder) => builder.AddSerilog())
-                .CreateWebHostBuilder()
+                .CreateWebHostBuider()
+                .UseContentRoot(AppContext.BaseDirectory)
                 .Build();
 
             var logger = webHost
@@ -130,8 +130,9 @@ namespace IdentityServer4.Contrib.AspNetCore.Testing.Tests
             var path = Path.Combine(AppContext.BaseDirectory, "Logs",
                 $"{Assembly.GetExecutingAssembly().GetName().Name}-{DateTime.UtcNow:yyyyMMdd}.log");
 
-            logger.LogInformation($"Logging to path {path} works!");
+            logger.LogError($"Logging to path {path} works!");
 
+            await Task.Delay(100);
             Assert.True(File.Exists(path));
         }
 
@@ -143,7 +144,7 @@ namespace IdentityServer4.Contrib.AspNetCore.Testing.Tests
             .WriteTo.Console(
                 outputTemplate:
                 "[{Timestamp:HH:mm:ss} {Level}] {SourceContext}{NewLine}{Message:lj}{NewLine}{Exception}{NewLine}",
-                theme: AnsiConsoleTheme.Literate)
+                theme: AnsiConsoleTheme.Literate, restrictedToMinimumLevel: LogEventLevel.Error)
             .CreateLogger();
     }
 }
